refactor: Step3 피드백 반영
